# functions/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Bet Chicago
host: localhost:5001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
  - application/vnd.contentful.management.v1+json
  - application/x-www-form-urlencoded
  - multipart/form-data
  - text/plain; charset=utf-8
  - image/jpeg
produces:
  - application/json
paths:
  /ap/uploadxml:
    x-swagger-router-controller: apupload
    post:
      description: upload ap xml article
      operationId: uploadAPArticleText
      tags:
        - AP Ingestion
      parameters:
        - name: filename
          in: query
          description: fileName of article
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /contentfulapi/paged:
    x-swagger-router-controller: contentfulapi
    post:
      description: paged results for cful
      operationId: contentfulPagedResults
      tags:
        - Contentful API
      parameters:
        - name: skip
          in: query
          description: records to skip (0 default)
          required: false
          type: string
        - name: pagesize
          in: query
          description: page size (10 default)
          required: false
          type: string
        - name: queryfilter
          in: query
          description: trending, section:id  (home / all articles default)
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /contentfulapi/articles:
    x-swagger-router-controller: contentfulapi
    post:
      description: paged results for cful
      operationId: contentfulArticleList
      tags:
        - Contentful API
      parameters:
        - name: idlist
          in: query
          description: articleids seperate by a space
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /contentfulapi/fetcharticlelistpage:
    x-swagger-router-controller: contentfulapi
    post:
      description: fetch article slugs for sitemap
      operationId: fetchArticleListPage
      tags:
        - Contentful API
      parameters:
        - name: skip
          in: query
          description: records to skip
          required: false
          type: string
        - name: pagesize
          in: query
          description: page size
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /golf/leaderboard:
    x-swagger-router-controller: golf
    post:
      description: update firedata from sports radar
      operationId: leaderboard
      tags:
        - Golf
      parameters:
        - name: tournamentId
          in: query
          description: id of tournament if not current
          required: false
          type: string
        - name: year
          in: query
          description: year :application.golfYearPGA is default
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /golf/tournament:
    x-swagger-router-controller: golf
    post:
      description: update firedata from sports radar
      operationId: tournament
      tags:
        - Golf
      parameters:
        - name: tournamentId
          in: query
          description: id of tournament :applicationConfig.golfTournament is default
          required: false
          type: string
        - name: year
          in: query
          description: year :application.golfYearPGA is default
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /componentslugs/updateComponentSlug:
    x-swagger-router-controller: componentslugs
    post:
      description: component slug
      operationId: updateComponentSlug
      tags:
        - ComponentSlug
      parameters:
        - name: root
          in: query
          description: Sport root url
          required: true
          type: string
        - name: tag
          in: query
          description: Sport sub menu
          required: true
          type: string
        - name: component
          in: query
          description: react component name
          required: true
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...